Added Tetris Game file
The game (so far) supports up to 4 players in parallel - implemented atm. is that only one figure can be moved while all others just move down steadily (also in multiplayer). It should serve as a basis to develop the required interfaces to interact with. The file includes pygame that allows for local testing of the functionality. 
For that - the game loop must be adapted. But for now, we should be able to work with that and can start to integrate the different parts.