New Method to avoid figure collision
Added a new method to avoid figure collisions that are controlled by players. I think it works however sometimes it seems like there can still be collisions. It could be a problem with pygame drawing that why we need to test it further once the 
web browser connection is set up.